refactor: moved component and its composables to a separate folder
